+++ src/FSharp.Formatting.ApiDocs/GenerateModel.fs
@@ -1,4 +1,4 @@
-namespace FSharp.Formatting.ApiDocs
+namespace rec FSharp.Formatting.ApiDocs
 
 open System
 open System.Reflection
@@ -254,8 +254,12 @@ type ApiDocMember(displayName: string, attributes: ApiDocAttribute list, entityU
     ApiDocAttribute.TryGetCustomOperationName(x.Attributes)
 
 /// Represents a type definition integrated with its associated documentation
-type ApiDocType(name, cat, url, comment, assembly, attributes, cases, fields, statParams, ctors, inst, stat) =
-    /// The name of the type
+type ApiDocTypeDefinition
+       (name, cat, url, comment, assembly: AssemblyName, attributes,
+        cases, fields, statParams, ctors, inst, stat, allInterfaces, baseType, abbreviatedType,
+        delegateSignature, symbol: FSharpEntity) =
+
+  /// The name of the type
   member x.Name : string = name
 
     /// The category of the type
@@ -268,7 +272,7 @@ type ApiDocType(name, cat, url, comment, assembly, attributes, cases, fields, st
   member x.Comment : ApiDocComment = comment
 
     /// The name of the type's assembly
-  member x.Assembly : AssemblyName = assembly
+  member x.Assembly = assembly
 
     /// The declared attributes of the type
   member x.Attributes : ApiDocAttribute list = attributes
@@ -284,6 +288,18 @@ type ApiDocType(name, cat, url, comment, assembly, attributes, cases, fields, st
 
     /// All members of the type
   member x.AllMembers : ApiDocMember list = List.concat [ ctors; inst; stat; cases; fields; statParams ]
+
+    /// All interfaces of the type, formatted
+  member x.AllInterfaces : string list = allInterfaces
+
+    /// The base type of the type, formatted
+  member x.BaseType : string option = baseType
+
+    /// If this is a type abbreviation, then the abbreviated type
+  member x.AbbreviatedType : string option = abbreviatedType
+
+    /// If this is a delegate, then e formatted signature
+  member x.DelegateSignature : string option = delegateSignature
 
     /// The constuctorsof the type
   member x.Constructors : ApiDocMember list = ctors
@@ -303,8 +319,13 @@ type ApiDocType(name, cat, url, comment, assembly, attributes, cases, fields, st
   member x.ObsoleteMessage =
     ApiDocAttribute.TryGetObsoleteMessage(x.Attributes)
 
+  /// The F# compiler symbol for the type definition
+  member x.Symbol = symbol
+
+
 /// Represents an F# module definition integrated with its associated documentation
-type ApiDocModule(name, cat, url, comment, assembly, attributes, modules, types, vals, exts, pats) =
+type ApiDocModule(name, cat, url, comment, assembly, attributes, modules, types, vals, exts, pats, symbol) =
+
     /// The name of the module
   member x.Name : string = name
 
@@ -330,7 +351,7 @@ type ApiDocModule(name, cat, url, comment, assembly, attributes, modules, types,
   member x.NestedModules : ApiDocModule list = modules
 
     /// All nested types
-  member x.NestedTypes : ApiDocType list = types
+  member x.NestedTypes : ApiDocTypeDefinition list = types
 
     /// Values and functions of the module
   member x.ValuesAndFuncs : ApiDocMember list = vals
@@ -350,6 +371,9 @@ type ApiDocModule(name, cat, url, comment, assembly, attributes, modules, types,
   member x.ObsoleteMessage =
     ApiDocAttribute.TryGetObsoleteMessage(x.Attributes)
 
+  /// The F# compiler symbol for the type definition
+  member x.Symbol = symbol
+
 /// Represents a namespace integrated with its associated documentation
 type ApiDocNamespace(name, mods, typs) =
     /// The name of the namespace
@@ -359,7 +383,7 @@ type ApiDocNamespace(name, mods, typs) =
   member x.Modules : ApiDocModule list = mods
 
     /// All types in the namespace
-  member x.Types : ApiDocType list = typs
+  member x.Types : ApiDocTypeDefinition list = typs
 
 /// Represents a group of assemblies integrated with its associated documentation
 type ApiDocAssemblyGroup(name: string, asms: AssemblyName list, nss: ApiDocNamespace list) =
@@ -378,7 +402,7 @@ type ApiDocModuleInfo(modul: ApiDocModule, asm: ApiDocAssemblyGroup, ns: ApiDocN
   member x.Module = modul
 
     /// The assembly group the module belongs to
-  member x.Assembly = asm
+  member x.AssemblyGroup = asm
 
     /// The namespace the module belongs to
   member x.Namespace = ns
@@ -386,16 +410,16 @@ type ApiDocModuleInfo(modul: ApiDocModule, asm: ApiDocAssemblyGroup, ns: ApiDocN
     /// The parent module, if any.
   member x.ParentModule = parent
 
-  member this.HasParentModule = this.ParentModule.IsSome
+  member x.HasParentModule = x.ParentModule.IsSome
 
 
 /// High-level information about a type definition
 type ApiDocTypeInfo(typ, asm, ns, modul) =
     /// The actual type
-  member x.Type : ApiDocType = typ
+  member x.Type : ApiDocTypeDefinition = typ
 
     /// The assembly group the type belongs to
-  member x.Assembly : ApiDocAssemblyGroup = asm
+  member x.AssemblyGroup : ApiDocAssemblyGroup = asm
 
     /// The namespace the type belongs to
   member x.Namespace : ApiDocNamespace = ns
@@ -403,7 +427,7 @@ type ApiDocTypeInfo(typ, asm, ns, modul) =
     /// The parent module, if any.
   member x.ParentModule : ApiDocModule option = modul
 
-  member this.HasParentModule = this.ParentModule.IsSome
+  member x.HasParentModule = x.ParentModule.IsSome
 
 
 module internal ValueReader =
@@ -433,18 +457,18 @@ module internal ValueReader =
       | _ -> None
 
     static member internal Create
-        (publicOnly, assembly, map, sourceFolderRepo, urlRangeHighlight, markDownComments, urlMap,
-         assemblyPath, compilerOptions, formatAgent ) =
+        (publicOnly, assembly, map, sourceFolderRepo, urlRangeHighlight, mdcomments, urlMap,
+         assemblyPath, fscoptions, formatAgent ) =
 
       { PublicOnly=publicOnly
         Assembly = assembly
         XmlMemberMap = map
-        MarkdownComments = markDownComments
+        MarkdownComments = mdcomments
         UrlMap = urlMap
         UrlRangeHighlight = urlRangeHighlight
         SourceFolderRepository = sourceFolderRepo
         AssemblyPath = assemblyPath
-        CompilerOptions = compilerOptions
+        CompilerOptions = fscoptions
         FormatAgent = formatAgent }
 
   let inline private getCompiledName (s : ^a when ^a :> FSharpSymbol) =
@@ -455,18 +479,18 @@ module internal ValueReader =
 
   let formatSourceLocation (urlRangeHighlight : Uri -> int -> int -> string) (sourceFolderRepo : (string * string) option) (location : range option) =
     location |> Option.bind (fun location ->
-        sourceFolderRepo |> Option.map (fun (baseFolder, repo) ->
-            let basePath = Uri(Path.GetFullPath(baseFolder)).ToString()
+        sourceFolderRepo |> Option.map (fun (sourceFolder, sourceRepo) ->
+            let sourceFolderPath = Uri(Path.GetFullPath(sourceFolder)).ToString()
             let docPath = Uri(Path.GetFullPath(location.FileName)).ToString()
 
             // Even though ignoring case might be wrong, we do that because
             // one path might be file:///C:\... and the other file:///c:\...  :-(
-            if not <| docPath.StartsWith(basePath, StringComparison.InvariantCultureIgnoreCase) then
-                Log.errorf "Current source file '%s' doesn't reside in source folder '%s'" docPath basePath
+            if not <| docPath.StartsWith(sourceFolderPath, StringComparison.InvariantCultureIgnoreCase) then
+                Log.verbf "Current source file '%s' doesn't reside in source folder '%s'" docPath sourceFolderPath
                 ""
             else
-                let relativePath = docPath.[basePath.Length..]
-                let uriBuilder = UriBuilder(repo)
+                let relativePath = docPath.[sourceFolderPath.Length..]
+                let uriBuilder = UriBuilder(sourceRepo)
                 uriBuilder.Path <- uriBuilder.Path + relativePath
                 urlRangeHighlight uriBuilder.Uri location.StartLine location.EndLine ) )
 
@@ -558,22 +582,25 @@ module internal ValueReader =
         formatTypeArgument typ.GenericParameter
     | _ -> "(type)"
 
-  let formatType typ =
-    formatTypeWithPrec 5 typ
+  let formatType (typ: FSharpType) =
+    formatTypeWithPrec 5 (FSharpType.Prettify typ)
 
   let isUnitType (ty: FSharpType) =
       ty.HasTypeDefinition && ty.TypeDefinition.XmlDocSig = "T:Microsoft.FSharp.Core.unit" 
 
-  let formatArgNameAndType i (arg:FSharpParameter) =
-    let nm =
-      match arg.Name with
+  let formatArgNameAndTypePair i (argName, argType) =
+    let argName =
+      match argName with
       | None ->
-          if isUnitType arg.Type then "()"
+          if isUnitType argType then "()"
           else "arg" + string i
       | Some nm -> nm
+    argName, argType
+
+  let formatArgNameAndType i (arg:FSharpParameter) =
+    let argName, argType  =  formatArgNameAndTypePair i (arg.Name, arg.Type)
     let isOptionalArg = arg.IsOptionalArg || hasAttrib<OptionalArgumentAttribute> arg.Attributes
-    let argName = if isOptionalArg then "?" + nm else nm
-    let argType = arg.Type
+    let argName = if isOptionalArg then "?" + argName else argName
     let argType =
         // Strip off the 'option' type for optional arguments
         if isOptionalArg && argType.HasTypeDefinition && argType.GenericArguments.Count = 1 then
@@ -587,6 +614,13 @@ module internal ValueReader =
     let argName, argType = formatArgNameAndType i arg
     if generateTypes then
       (match arg.Name with None -> "" | Some argName -> argName + ": ") +
+      formatTypeWithPrec 2 argType
+    else argName
+
+  let formatArgNameAndTypePairUsage generateTypes i (argName0, argType) =
+    let argName, argType = formatArgNameAndTypePair i (argName0, argType)
+    if generateTypes then
+      (match argName0 with None -> "" | Some argName -> argName + ": ") +
       formatTypeWithPrec 2 argType
     else argName
 
@@ -611,6 +645,9 @@ module internal ValueReader =
     | Some loc -> Some loc
     | None -> symbol.DeclarationLocation
 
+  let formatDelegateSignature (typ: FSharpDelegateSignature) =
+    (typ.DelegateArguments |> List.ofSeq |> List.mapi (formatArgNameAndTypePairUsage true) |> String.concat " * ") + " -> " + formatType typ.DelegateReturnType 
+
   let readAttribute (attribute: FSharpAttribute) =
     let name = attribute.AttributeType.DisplayName
     let fullName = attribute.AttributeType.FullName
@@ -624,7 +661,6 @@ module internal ValueReader =
     |> Seq.toList
 
   let readMemberOrVal (ctx:ReadingContext) (v:FSharpMemberOrFunctionOrValue) =
-    // we calculate this early just in case this fails with an FCS error.
     let requireQualifiedAccess =
         hasAttrib<RequireQualifiedAccessAttribute> v.ApparentEnclosingEntity.Attributes
 
@@ -637,8 +673,13 @@ module internal ValueReader =
           if String.IsNullOrWhiteSpace(s) then ""
           elif s.StartsWith("(") then s
           // curried arguments should not have brackets, see https://github.com/fsprojects/FSharp.Formatting/issues/472
-          elif argInfos.Length > 1 then " " + s
-          else sprintf "(%s)" s)
+          // single argument F# function values should not have brackets if they are not tuples 
+          // single argument F# members should not have brackets if they are lower case
+          elif argInfos.Length > 1 ||
+               (s <> "()" && not (s.Contains(",")) && (not v.IsMember || (Char.IsLower(v.DisplayName.[0])))) then
+              " " + s
+          else
+              sprintf "(%s)" s)
 
       match v.IsMember, v.IsInstanceMember, v.LogicalName, v.DisplayName with
       // Constructors 
@@ -689,7 +730,9 @@ module internal ValueReader =
     let retTypeText = defaultArg (retType |> Option.map formatType) "unit"
 
     let returnTooltip =
-       match retType with None -> None | Some retType -> if isUnitType retType then None else Some retTypeText 
+       match retType with
+       | None -> None
+       | Some retType -> if isUnitType retType then None else Some retTypeText 
 
     let signatureTooltip =
       match argInfos with
@@ -879,7 +922,7 @@ module internal Reader =
           ()
        ) (e.Nodes())
    readElement doc
-   full.Append("</br>") |> ignore
+   //full.Append("</br>") |> ignore
 
    let summaries = doc.Descendants(XName.Get "summary")
    summaries |> Seq.iteri (fun id e ->
@@ -1060,7 +1103,7 @@ module internal Reader =
             let doc =
                 Literate.ParseMarkdownString
                   ( text, path=Path.Combine(ctx.AssemblyPath, "docs.fsx"),
-                    formatAgent=ctx.FormatAgent, compilerOptions=ctx.CompilerOptions )
+                    formatAgent=ctx.FormatAgent, fscoptions=ctx.CompilerOptions )
                   |> (addMissingLinkToTypes ctx)
             cmds :> IDictionary<_, _>, readMarkdownComment doc
           else
@@ -1114,8 +1157,7 @@ module internal Reader =
                 else ""
 
             let paramList =
-                if memb.CurriedParameterGroups.Count > 0 && memb.CurriedParameterGroups.[0].Count > 0
-                then
+                if memb.CurriedParameterGroups.Count > 0 && memb.CurriedParameterGroups.[0].Count > 0 then
                     let head = memb.CurriedParameterGroups.[0]
                     let paramTypeList =
                         head
@@ -1128,7 +1170,7 @@ module internal Reader =
                 else ""
             sprintf "%s%s%s" name typeargs paramList
           with exn ->
-            Log.errorf "Error while building member-name for %s because: %s" memb.FullName exn.Message
+            printfn "Error while building member-name for %s because: %s" memb.FullName exn.Message
             Log.verbf "Full Exception details of previous message: %O" exn
             memb.CompiledName
         match (memb.DeclaringEntity.Value.TryFullName) with
@@ -1291,7 +1333,7 @@ module internal Reader =
                   | None -> ass.QualifiedName
                 with _ -> "unknown"
               sprintf "unknown, part of %s" part
-        Log.errorf "Could not read comments from entity '%s': %O" name e
+        printfn "Could not read comments from entity '%s': %O" name e
         None
 
   let checkAccess ctx (access: FSharpAccessibility) =
@@ -1416,17 +1458,10 @@ module internal Reader =
           |> List.partition (fun v -> v.IsInstanceMember)
       let cvals, svals = svals |> List.partition (fun v -> v.CompiledName = ".ctor")
 
-      (*
-      // Base types?
-      let iimpls =
-        if ( not typ.IsAbbreviation && not typ.HasAssemblyCodeRepresentation &&
-             typ.ReflectionType.IsInterface) then [] else typ.Implements |> List.ofSeq
-      // TODO: layout base type in some way
-      if not iimpls.IsEmpty then
-        newTable1 hFile "Interfaces" 40 "Type"  (fun () ->
-          iimpls |> List.iter (fun i ->
-              newEntry1 hFile ("<pre>"+outputL widthVal (layoutType denv i)+"</pre>")))
-      *)
+      let baseType = typ.BaseType |> Option.map formatType
+      let allInterfaces = [ for i in typ.AllInterfaces -> formatType i ]
+      let abbreviatedType = if typ.IsFSharpAbbreviation then Some (formatType typ.AbbreviatedType) else None 
+      let delegateSignature = if typ.IsDelegate then Some (typ.DisplayName + formatDelegateSignature typ.FSharpDelegateSignature) else None 
 
       let name = readTypeName typ
       let cases = readUnionCases ctx entityUrl typ
@@ -1439,7 +1474,8 @@ module internal Reader =
       let inst = readAllMembers ctx entityUrl ApiDocMemberKind.InstanceMember ivals
       let stat = readAllMembers ctx entityUrl ApiDocMemberKind.StaticMember svals
 
-      ApiDocType (name, cat, entityUrl, comment, ctx.Assembly, attrs, cases, fields, statParams, ctors, inst, stat ))
+      ApiDocTypeDefinition (name, cat, entityUrl, comment, ctx.Assembly, attrs, cases, fields, statParams, ctors,
+         inst, stat, allInterfaces, baseType, abbreviatedType, delegateSignature, typ))
 
   and readModule (ctx:ReadingContext) (modul:FSharpEntity) =
     readCommentsInto modul ctx modul.XmlDocSig (fun cat cmd comment ->
@@ -1455,8 +1491,7 @@ module internal Reader =
 
       ApiDocModule
         ( modul.DisplayName, cat, entityUrl, comment, ctx.Assembly, attrs,
-          modules, types,
-          vals, exts, pats ))
+          modules, types, vals, exts, pats, modul ))
 
   // ----------------------------------------------------------------------------------------------
   // Reading namespace and assembly details
@@ -1466,7 +1501,7 @@ module internal Reader =
     let modules, types = readModulesAndTypes ctx entities
     ApiDocNamespace(ns, modules, types)
 
-  let readAssembly (assembly:FSharpAssembly, publicOnly, xmlFile:string, sourceFolderRepo, urlRangeHighlight, markDownComments, urlMap, codeFormatCompilerArgs) =
+  let readAssembly (assembly:FSharpAssembly, publicOnly, xmlFile:string, sourceFolderRepo, urlRangeHighlight, mdcomments, urlMap, codeFormatCompilerArgs) =
     let assemblyName = AssemblyName(assembly.QualifiedName)
 
     // Read in the supplied XML file, map its name attributes to document text
@@ -1496,7 +1531,7 @@ module internal Reader =
     let ctx =
       ReadingContext.Create
         (publicOnly, assemblyName, xmlMemberMap, sourceFolderRepo, urlRangeHighlight,
-         markDownComments, urlMap, asmPath, codeFormatCompilerArgs, formatAgent)
+         mdcomments, urlMap, asmPath, codeFormatCompilerArgs, formatAgent)
 
     //
     let namespaces =
@@ -1519,7 +1554,7 @@ type ApiDocsModel =
     CollectionRootUrl  : string
   }
 
-  static member Generate(dllFiles: seq<string>, parameters, xmlFile, sourceRepo, sourceFolder, publicOnly, libDirs, otherFlags, markDownComments, urlRangeHighlight, rootUrl) =
+  static member Generate(dllFiles: seq<string>, parameters, xmlFile, sourceRepo, sourceFolder, publicOnly, libDirs, otherFlags, mdcomments, urlRangeHighlight, rootUrl) =
     let (@@) a b = Path.Combine(a, b)
     let parameters = defaultArg parameters []
     let props = [ "Properties", dict parameters ]
@@ -1540,7 +1575,7 @@ type ApiDocsModel =
             Log.warnf "Repository url should be specified along with source folder."
             None
         | _, Some repo ->
-            Log.warnf "Source folder should be specified along with repository url."
+            Log.warnf "Repository url should be specified along with source folder."
             None
         | _ -> None
 
@@ -1557,7 +1592,7 @@ type ApiDocsModel =
                 let bytes = File.ReadAllBytes(file)
                 Some(System.Reflection.Assembly.Load(bytes))
             with e ->
-              Log.errorf "Couldn't load Assembly\n%s\n%s" e.Message e.StackTrace
+              printfn "Couldn't load Assembly\n%s\n%s" e.Message e.StackTrace
               None
           else None )
       defaultArg asmOpt null
@@ -1611,11 +1646,11 @@ type ApiDocsModel =
           | Some xmlFile ->
             Reader.readAssembly
               (asm, publicOnly, xmlFile, sourceFolderRepo, urlRangeHighlight,
-               defaultArg markDownComments true, urlMap, codeFormatCompilerArgs )
+               defaultArg mdcomments true, urlMap, codeFormatCompilerArgs )
             |> Some)
 
     // Get the name - either from parameters, or name of the assembly (if there is just one)
-    let name =
+    let groupName =
       let projName = parameters |> List.tryFind (fun (k, v) -> k = "project-name") |> Option.map snd
       match assemblies, projName with
       | _, Some name -> name
@@ -1635,32 +1670,32 @@ type ApiDocsModel =
           if mods.Length + typs.Length > 0 then
               ApiDocNamespace(name, mods, typs) ]
 
-    let asm = ApiDocAssemblyGroup(name, List.map fst assemblies, namespaces |> List.sortBy (fun ns -> ns.Name))
+    let asmGroup = ApiDocAssemblyGroup(groupName, List.map fst assemblies, namespaces |> List.sortBy (fun ns -> ns.Name))
 
     let rec nestedModules ns parent (modul:ApiDocModule) =
       seq {
-        yield ApiDocModuleInfo(modul, asm, ns, parent)
+        yield ApiDocModuleInfo(modul, asmGroup, ns, parent)
         for n in modul.NestedModules do yield! nestedModules ns (Some modul) n
       }
 
     let moduleInfos =
-      [ for ns in asm.Namespaces do
+      [ for ns in asmGroup.Namespaces do
           for n in ns.Modules do yield! nestedModules ns None n ]
 
     let createType ns modul typ =
-        ApiDocTypeInfo(typ, asm, ns, modul)
+        ApiDocTypeInfo(typ, asmGroup, ns, modul)
 
     let rec nestedTypes ns (modul:ApiDocModule) = seq {
       yield! (modul.NestedTypes |> List.map (createType ns (Some modul) ))
       for n in modul.NestedModules do yield! nestedTypes ns n }
 
     let typesInfos =
-      [ for ns in asm.Namespaces do
+      [ for ns in asmGroup.Namespaces do
           for n in ns.Modules do yield! nestedTypes ns n
           yield! (ns.Types |> List.map (createType ns None ))  ]
 
     {
-      AssemblyGroup = asm
+      AssemblyGroup = asmGroup
       ModuleInfos = moduleInfos
       TypesInfos = typesInfos
       Properties = props
@@ -1686,8 +1721,10 @@ type DocComment = class end
 type Module = class end
 [<Obsolete("Renamed to ApiDocModuleInfo", true)>]
 type ModuleInfo = class end
-[<Obsolete("Renamed to ApiDocType", true)>]
+[<Obsolete("Renamed to ApiDocTypeDefinition", true)>]
 type Type = class end
+[<Obsolete("Renamed to ApiDocTypeDefinition", true)>]
+type ApiDocType = class end
 [<Obsolete("Renamed to ApiDocTypeInfo", true)>]
 type TypeInfo = class end